feat: DB 적용 - JdbcTemplate

+++ src/main/java/nextstep/presentation/dto/ReservationResponse.java
@@ -2,7 +2,7 @@ package nextstep.presentation.dto;
 
 public class ReservationResponse {
 
-    private Integer id;
+    private Long id;
     private String date;
     private String time;
     private String name;
@@ -10,14 +10,14 @@ public class ReservationResponse {
     private ReservationResponse() {
     }
 
-    public ReservationResponse(Integer id, String date, String time, String name) {
+    public ReservationResponse(Long id, String date, String time, String name) {
         this.id = id;
         this.date = date;
         this.time = time;
         this.name = name;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
